Extracted common logic to AndroidHttpService
